# Speed up and avoid kwarg hash alloc in Time.now