--- a/benchmark/time_now.yml
+++ b/benchmark/time_now.yml
@@ -1,0 +1,3 @@
+benchmark:
+  - 'Time.now'
+  - 'Time.now(in: "+09:00")'